fixed some paths

+++ graph_embeddings/perturbations.py
@@ -18,17 +18,13 @@ from .cython_metric import custom_hake_pairwise_distances
 CONFIG = configparser.ConfigParser()
 CONFIG.read('config.ini')
 
-EMBEDDINGS_DIR = Path(CONFIG['EMBEDDING FILES']['trained_embeddings'])
-NODE_EMBEDDINGS_FILE = EMBEDDINGS_DIR / Path(
-    CONFIG['EMBEDDING FILES']['node_embeddings_file'])
-REL_EMBEDDINGS_FILE = EMBEDDINGS_DIR / Path(
-    CONFIG['EMBEDDING FILES']['rel_embeddings_file'])
-LAMBDA_FILE = EMBEDDINGS_DIR / Path(CONFIG['EMBEDDING FILES']['lambda_file'])
+EMBEDDINGS_DIR = Path(CONFIG['EMBEDDING FILES']['embeddings_dir'])
+NODE_EMBEDDINGS_FILE = Path(CONFIG['EMBEDDING FILES']['node_embeddings_file'])
+REL_EMBEDDINGS_FILE = Path(CONFIG['EMBEDDING FILES']['rel_embeddings_file'])
+LAMBDA_FILE = Path(CONFIG['EMBEDDING FILES']['lambda_file'])
 
-DENSE_DISTS_MAT_FILE = EMBEDDINGS_DIR / CONFIG['EMBEDDING FILES'][
-    'dense_dists_mat_file']
-DENSE_PROBS_MAT_FILE = EMBEDDINGS_DIR / CONFIG['EMBEDDING FILES'][
-    'dense_probs_mat_file']
+DENSE_DISTS_MAT_FILE = CONFIG['EMBEDDING FILES']['dense_dists_mat_file']
+DENSE_PROBS_MAT_FILE = CONFIG['EMBEDDING FILES']['dense_probs_mat_file']
 
 
 def translate_hake_embedding(node_embedding, rel_embedding, lam=1.0):
@@ -83,9 +79,9 @@ def get_softmax_probs(dist_mats, ratio=0.8, thresh_1=5.0, thresh_2=6.0):
 
             remainder = (1 - ratio) / 2.0
 
-            row_softmax_probs[s1_indices] *= remainder / s1
-            row_softmax_probs[s2_indices] *= ratio / s2
-            row_softmax_probs[s3_indices] *= remainder / s3
+            row_softmax_probs[s1_indices] *= 0.99 / s1
+            row_softmax_probs[s2_indices] *= 0.005 / s2
+            row_softmax_probs[s3_indices] *= 0.005 / s3
 
             row_softmax_probs = row_softmax_probs / np.sum(row_softmax_probs)
             softmax_probs[i] = row_softmax_probs
@@ -98,9 +94,9 @@ def get_softmax_probs(dist_mats, ratio=0.8, thresh_1=5.0, thresh_2=6.0):
     return softmax_mat
 
 
-def distance_kernel(x, l, t, v, u, k_max, k_min):
-    b = 10**(np.log10(float(k_max) / k_min) / (l * (u - v)))
-    a = float(k_max) / b**(-l * (t + v))
+def distance_kernel(x, seq_len, threshold, v, u, k_max, k_min):
+    b = 10**(np.log10(float(k_max) / k_min) / (seq_len * (u - v)))
+    a = float(k_max) / b**(-seq_len * (threshold + v))
 
     k = a * b**(-x)
     k = np.minimum(np.ones_like(k), k)
@@ -108,31 +104,38 @@ def distance_kernel(x, l, t, v, u, k_max, k_min):
     return k
 
 
-def save_embedding_distances_and_probs(output_dir):
-    
-    with open(NODE_EMBEDDINGS_FILE, 'rb') as f:
+def save_embedding_distances_and_probs(output_dir, use_cached_dists=True):
+
+    with open(Path(output_dir) / NODE_EMBEDDINGS_FILE, 'rb') as f:
         node_embeddings = np.load(f)
 
-    with open(REL_EMBEDDINGS_FILE, 'rb') as f:
+    with open(Path(output_dir) / REL_EMBEDDINGS_FILE, 'rb') as f:
         rel_embeddings = np.load(f)
 
-    with open(LAMBDA_FILE, 'rb') as f:
+    with open(Path(output_dir) / LAMBDA_FILE, 'rb') as f:
         lam_dict = json.load(f)
         lam = lam_dict['lambda']
         lam2 = lam_dict['lambda2']
 
-    dist_mats = get_embedding_distances(rel_embeddings, node_embeddings, lam,
-                                        lam2)
+    if use_cached_dists and Path.exists(
+            Path(output_dir) / DENSE_DISTS_MAT_FILE):
+        dist_mats = np.load(Path(output_dir) / DENSE_DISTS_MAT_FILE)
+    else:
+        dist_mats = get_embedding_distances(rel_embeddings, node_embeddings,
+                                            lam, lam2)
     softmax_probs = get_softmax_probs(dist_mats,
-                                      ratio=0.8,
+                                      ratio=0.50,
                                       thresh_1=5.0,
                                       thresh_2=6.0)
 
     # Save dense matrices
-    with open(DENSE_DISTS_MAT_FILE, 'wb') as f:
-        np.save(f, dist_mats)
 
-    with open(DENSE_PROBS_MAT_FILE, 'wb') as f:
+    if not (use_cached_dists
+            and Path.exists(Path(output_dir) / DENSE_DISTS_MAT_FILE)):
+        with open(Path(output_dir) / DENSE_DISTS_MAT_FILE, 'wb') as f:
+            np.save(f, dist_mats)
+
+    with open(Path(output_dir) / DENSE_PROBS_MAT_FILE, 'wb') as f:
         np.save(f, softmax_probs)
 
 
@@ -142,7 +145,6 @@ if __name__ == '__main__':
         "Calculate distances and replacement probabilities between node embeddings."
     )
     parser.add_argument('output_dir',
-                        '-o',
                         help='Directory containing model output.')
     args = parser.parse_args()
